Implement service.RunWithSighandler
